display_control: Make display_name include monitor name, serial number, etc
Currently the display_name() function uses only the 'id' field of the
ddc_hi::DisplayInfo struct, plus a possible numeric suffix to disambiguate
otherwise-identical display names.  At least on Linux, this leaves a lot of
useful information out of the display name, because for the i2c-dev backend
to ddc_hi the 'id' string is just the decimal representation of the
(major,minor) of the i2c device node (e.g., '22789' == 0x5905 == device
(59,05) == /dev/i2c-5).

There is usually more usefully identifying information in the DisplayInfo
structure, including the 3-character manufacturer ID, the model name, and
the monitor serial number. Append this information to 'id' if it is
present. The result is that instead of logging like
  Display '22790' is currently set to Custom(0x1111)
you get
  Display '22790 DEL DELL U2719D DT86VS2' is currently set to Custom(0x1111)
and you can also write monitor_id config file settings that match
against the monitor model or serial number.